Start of a Warnings strategy; Set plugin type
A variety of changes to the CMakeLists

First, set the pugin type to a synth which needs midi input.
Ooops!

Second, start to work through all the warnings in the codebase

1. Turns on warnings-are-errors on all paltforms and then
2. Supresses a whole bunch of warnings, some of them rather
   legitimate, in the CMake file

This means at least a build has readable output, but also
defines the book of work we need to understake to finish #16