Don't return temporaries into a StringRef
The pattern

StringRef foo() { return String( "bar" ); }

can result in the string being freed before the ref is resolved.
Address sanitizer shows this as a problem, as described in #52.
I'm not sure if this fix is the right one or if we should instead
change the API to String rather than StringRef, and especially
the dynamic string one is iffy, but this at least squashes the
memory problem the sanitizer showed in single instance.