Repository layout
Monique has moved away from large commits across the repo,
the later this is done the more history will be blurred.
So here is my layout suggestion.
git blame and the like still works, but githubs History feature as well as some
"one-click-timeline-solutions" of some IDEs will be broken (showing this as the first commit).

The problem I have with this is: If I commit this, GH will display me as the one and only author
of all the files, which is far from the truth. Does anyone know a trick to prevent that from happening?
Otherwise I'd happily close this and ask @TommmyArn to do the commit (feel free to copy/paste of course).

Closes #30