Repository layout
Monique has moved away from large commits across the repo,
the later this is done the more history will be blurred.
So here is my layout suggestion.
git blame and the like still works, but githubs History feature as well as some
"one-click-timeline-solutions" of some IDEs will be broken (showing this as the first commit).

I changed the committer name to SST so Github doesn't incorrectly display me as the only author
of all the files (let's see if this works)

Closes #30